create npm package?
