+++ README.md
@@ -1,0 +1,6 @@
+Slabtype
+========
+
+A port of [Erik Loyer's slabtype algorithm][algo] to Javascript & Canvas.
+
+[algo]: http://erikloyer.com/index.php/blog/the_slabtype_algorithm_part_1_background/
