+++ index.html
@@ -99,7 +99,7 @@
         <canvas class="text">This claim contains exactly 100 characters & no more, which is probably too many characters for us. Plus maybe a few more, just to make sure it's too darn big.</canvas>
     </div>
 
-    <script src="slabtype.js"></script>
+    <script src="dist/slabtype.js"></script>
     <script>
         var canvasEls = document.querySelectorAll('.text');
         Array.prototype.slice.call(canvasEls).forEach(function(el) {
