+++ package.json
@@ -1,0 +1,26 @@
+{
+  "name": "slabtype",
+  "version": "1.0.0",
+  "description": "Fancy text layout in a box",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/mccutchen/slabtype.git"
+  },
+  "keywords": [
+    "client",
+    "canvas",
+    "typography",
+    "layout",
+    "font"
+  ],
+  "author": "Will McCutchen <will@mccutch.org>",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/mccutchen/slabtype/issues"
+  },
+  "homepage": "https://github.com/mccutchen/slabtype",
+  "devDependencies": {
+    "browserify": "6.1.0"
+  },
+  "main": "src/slabtype.js"
+}
